+++ packages/xchain-client/src/types.ts
@@ -4,7 +4,6 @@ export type Address = string
 
 export enum Network {
   Mainnet = 'mainnet',
-  Stagenet = 'stagenet',
   Testnet = 'testnet',
 }