+++ packages/xchain-polkadot/src/util.ts
@@ -18,7 +18,6 @@ export const isSuccess = (response: { code: number }): boolean => !response.code
 export const getDecimal = (network: Network): number => {
   switch (network) {
     case Network.Mainnet:
-    case Network.Stagenet:
       return 10
     case Network.Testnet:
       return 12
@@ -46,7 +45,6 @@ export const getDefaultFees = (network: Network): Fees => {
 export const getPrefix = (network: Network) => {
   switch (network) {
     case Network.Mainnet:
-    case Network.Stagenet:
       return '1'
     case Network.Testnet:
       return '5'