+++ packages/xchain-terra/__e2e__/client.e2e.ts
@@ -1,6 +1,6 @@
-import { /*Balance,*/ Network, Tx } from '@xchainjs/xchain-client'
-import { Chain } from '@xchainjs/xchain-util'
-//import { Chain, baseAmount, baseToAsset } from '@xchainjs/xchain-util'
+import { /*Balance,*/ Network, Tx } from '@thorswap-lib/xchain-client'
+import { Chain } from '@thorswap-lib/xchain-util'
+//import { Chain, baseAmount, baseToAsset } from '@thorswap-lib/xchain-util'
 
 import { Client } from '../src'